removed P15CIB and fixed premasking in SPT. CIB still remains nan open question
