+++ xcell/mappers/mapper_SPT.py
@@ -20,6 +20,7 @@ class MapperSPT(MapperPlanckBase):
         self.ps_mask_mode = config.get('ps_mask_mode', ['default'])
         # Fix rotation from Planck's default
         self.rot = self._get_rotator('C')
+        self.pre_mask_ps = config.get("pre_mask_ps", True)
 
     def _get_hm_maps(self):
         if self.hm1_map is None:
@@ -27,6 +28,14 @@ class MapperSPT(MapperPlanckBase):
             hm1_map[hm1_map == hp.UNSEEN] = 0.0
             hm1_map[np.isnan(hm1_map)] = 0.0
             hm1_map = rotate_map(hm1_map, self.rot)
+            if self.pre_mask_ps is True
+                if self.file_ps_mask is not None
+                    ps_mask = self._get_ps_mask()
+                    hm1_map *= ps_mask
+                else
+                    NotImplementedError("""Tried to pre-mask point
+                                        sources but couldn't find 
+                                        file_ps_mask""")
             self.hm1_map = [hp.ud_grade(hm1_map,
                             nside_out=self.nside)]
         if self.hm2_map is None:
@@ -34,6 +43,14 @@ class MapperSPT(MapperPlanckBase):
             hm2_map[hm2_map == hp.UNSEEN] = 0.0
             hm2_map[np.isnan(hm2_map)] = 0.0
             hm2_map = rotate_map(hm2_map, self.rot)
+            if self.pre_mask_ps is True
+                if self.file_ps_mask is not None
+                    ps_mask = self._get_ps_mask()
+                    hm2_map *= ps_mask
+                else
+                    NotImplementedError("""Tried to pre-mask point
+                                        sources but couldn't find 
+                                        file_ps_mask""")
             self.hm2_map = [hp.ud_grade(hm2_map,
                             nside_out=self.nside)]
         return self.hm1_map, self.hm2_map
